+++ DataAccess/DataAccess.csproj
@@ -54,8 +54,6 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ContactForm\ContactFormRepo.cs" />
-    <Compile Include="ContactForm\IContactFormRepo.cs" />
     <Compile Include="ContentRepo\ContentRepo.cs" />
     <Compile Include="Context\AppsContext.cs" />
     <Compile Include="Entities\COREContent.cs" />
@@ -68,7 +66,6 @@
     <Compile Include="Entities\MenuItemType.cs" />
     <Compile Include="ImageItem\ImageItemRepo.cs" />
     <Compile Include="MenuItem\MenuItemRepo.cs" />
-    <Compile Include="Models\ContactFormModel.cs" />
     <Compile Include="Models\MenuItemFormModel.cs" />
     <Compile Include="Models\MenuItemModel.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
