+++ RestaurantPlay2/Areas/ContactForm/Views/Contact/Index.cshtml
@@ -47,9 +47,9 @@
             @Html.DisplayFor(modelItem => item.Comments)
         </td>
         <td>
-            @Html.ActionLink("Edit", "Edit", new { id=item.EnquiryId }) |
-            @Html.ActionLink("Details", "Details", new { id=item.EnquiryId }) |
-            @Html.ActionLink("Delete", "Delete", new { id=item.EnquiryId })
+            @Html.ActionLink("Edit", "Edit", new { id=item.Id }) |
+            @Html.ActionLink("Details", "Details", new { id=item.Id }) |
+            @Html.ActionLink("Delete", "Delete", new { id=item.Id })
         </td>
     </tr>
 }
