+++ RestaurantPlay2/Areas/ContactForm/Views/Contact/Success.cshtml
@@ -1,0 +1,12 @@
+﻿<div class="container-fluid bg-2 text-center">
+    <h3 class="margin">What now??</h3>
+    <p>We will be in touch shortly!!</p>
+</div>
+<style>
+    .margin { margin-bottom: 45px; }
+    .bg-1 { background-color: #1abc9c; /* Green */ color: #ffffff; }
+    .bg-2 { background-color: #474e5d; /* Dark Blue */ color: #ffffff; }
+    .bg-3 { background-color: #ffffff; /* White */ color: #555555; }
+    .bg-4 { background-color: #2f2f2f; /* Black Gray */ color: #fff; }
+    .container-fluid { padding-top: 70px; padding-bottom: 70px; }
+</style>
