+++ RestaurantPlay2/Content/Site.css
@@ -91,3 +91,27 @@ img.card-img-top {
     background-color: rgba(0, 0, 0, 0.60);
     ;
 }
+
+/*This is for the modal on the edit Enquiry Page*/
+#myImg { border-radius: 5px; cursor: pointer; transition: 0.3s; }
+
+#myImg:hover { opacity: 0.7; }
+.modal { display: none; /* Hidden by default */ position: fixed; /* Stay in place */ z-index: 1; /* Sit on top */ padding-top: 100px; /* Location of the box */ left: 0; top: 0; width: 100%; /* Full width */ height: 100%; /* Full height */ overflow: auto; /* Enable scroll if needed */ background-color: rgb(0,0,0); /* Fallback color */ background-color: rgba(0,0,0,0.9); /* Black w/ opacity */ }
+.modal-content { margin: auto; display: block; width: 80%; max-width: 700px; }
+#caption { margin: auto; display: block; width: 80%; max-width: 700px; text-align: center; color: #ccc; padding: 10px 0; height: 150px; }
+.modal-content, #caption { -webkit-animation-name: zoom; -webkit-animation-duration: 0.6s; animation-name: zoom; animation-duration: 0.6s; }
+@-webkit-keyframes zoom {
+    from { -webkit-transform: scale(0) }
+    to { -webkit-transform: scale(1) }
+}
+@keyframes zoom {
+    from { transform: scale(0) }
+    to { transform: scale(1) }
+}
+.close { position: absolute; top: 65px; right: 35px; color: #f1f1f1; font-size: 40px; font-weight: bold; transition: 0.3s; }
+.close:hover,
+.close:focus { color: #bbb; text-decoration: none; cursor: pointer; }
+@media only screen and (max-width: 700px) {
+    .modal-content { width: 100%; }
+}
+.modal-content div { padding: 20px; }
