+++ RestaurantPlay2/RestaurantPlay2.csproj
@@ -221,8 +221,7 @@
     <Compile Include="Areas\Blogger\BloggerAreaRegistration.cs" />
     <Compile Include="Areas\Blogger\Controllers\BloggerController.cs" />
     <Compile Include="Areas\ContactForm\ContactFormAreaRegistration.cs" />
-    <Compile Include="Areas\ContactForm\Controllers\EnquiryController.cs" />
-    <Compile Include="Areas\ContactForm\ViewModels\ContactFormViewModel.cs" />
+    <Compile Include="Areas\ContactForm\Controllers\ContactController.cs" />
     <Compile Include="Areas\ImageItems\BusinessLogic\ImageCards\ImageCardLogic.cs" />
     <Compile Include="Areas\ImageItems\BusinessLogic\ImageCarousel\ImageCarouselLogic.cs" />
     <Compile Include="Areas\ImageItems\BusinessLogic\Misc\ImageUtils.cs" />
@@ -317,11 +316,13 @@
     <Content Include="Areas\ContactForm\Views\web.config" />
     <Content Include="Areas\ContactForm\Views\_ViewStart.cshtml" />
     <Content Include="Areas\ContactForm\Views\Shared\_Layout.cshtml" />
-    <Content Include="Areas\ContactForm\Views\Enquiry\Create.cshtml" />
-    <Content Include="Areas\ContactForm\Views\Enquiry\Delete.cshtml" />
-    <Content Include="Areas\ContactForm\Views\Enquiry\Details.cshtml" />
-    <Content Include="Areas\ContactForm\Views\Enquiry\Edit.cshtml" />
-    <Content Include="Areas\ContactForm\Views\Enquiry\Index.cshtml" />
+    <Content Include="Areas\ContactForm\Views\Contact\Create.cshtml" />
+    <Content Include="Areas\ContactForm\Views\Contact\Delete.cshtml" />
+    <Content Include="Areas\ContactForm\Views\Contact\Details.cshtml" />
+    <Content Include="Areas\ContactForm\Views\Contact\Edit.cshtml" />
+    <Content Include="Areas\ContactForm\Views\Contact\Index.cshtml" />
+    <Content Include="Areas\ContactForm\Views\Contact\Contact.cshtml" />
+    <Content Include="Areas\ContactForm\Views\Contact\Success.cshtml" />
     <None Include="Scripts\jquery-1.10.2.intellisense.js" />
     <Content Include="Scripts\CustomScripts\CustomScripts.js" />
     <Content Include="Scripts\jquery-1.10.2.js" />
